Use node-pre-gyp to build native assets
Since we need to install a native component using node-gyp, currently anybody running "npm install"
must have a compiler installed. This is routine on Linux, but an annoyance on Windows. It also
makes it difficult to build for a different architecture than your current host.

This change introduces node-pre-gyp, which makes it possible to download a prebuilt binary from a
specified host. For this to work for the end-user we need to make two changes:

1. Upon a Github release, use node-pre-gyp to create binaries for Linux, MacOS and Windows
across all supported NodeJS versions. We then augment the release with these binaries.

2. Modify package.json to support downloading binaries from Github on package installation.
This assumes Github releases named after their version number (this can be changed in the binaries
section of package.json).