Fatal error if OnOOMError is called more than once
Creating a heapdump forces a garbage collection, which can then trigger
the OOM flow again, which then calls OnOOMError which then creates a
heapdump which forces GC which triggers OOM which... This recursive loop
is bad for several reasons, not the least of which being if addTimestamp
is true it will keep appending the timestamp to the filename, even past
the end of the allotted 256 chars. Rather than let this happen let's
instead treat it as an error condition and bail.