Raise memory limit when capturing heapdump
As noted in the last commit, capturing a heap snapshot forces a garbage
collection which can recursively trigger the OOM flow. To prevent this
we now raise the heap size limit while capturing this last dump. As far
as I can tell this is safe, even in memory-constrained environments,
because by the time we get to this section the JS execution has already
been halted, so the JS heap won't grow any further.